services/technical-seo-analyzer/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# HTTP & Scraping
httpx==0.25.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# AI & NLP
anthropic==0.7.0
textstat==0.7.3

# Utilities
python-dotenv==1.0.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Caching (optional)
redis==5.0.1
